# trilli-serve/system/apikeys/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trilli API",
    "version": "1.0.0",
    "description": "Programmatic access to your Trilli account — list, upload, download, and organize files and folders. The API is a separate interface to the same account and obeys the same plan limits (storage, max file size, and monthly transfer). Available on plans that include API access (Plus and Infinity)."
  },
  "servers": [
    { "url": "/api/v1", "description": "Trilli API v1 (relative to your Trilli host)" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/folders": {
      "get": {
        "tags": ["Folders"],
        "summary": "List folders",
        "description": "Lists subfolders of a folder. Omit `parent_id` to list top-level folders.",
        "parameters": [
          { "name": "parent_id", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" }, "description": "Parent folder. Omit for top-level folders." }
        ],
        "responses": {
          "200": { "description": "An array of folders", "content": { "application/json": { "schema": { "type": "object", "properties": { "folders": { "type": "array", "items": { "$ref": "#/components/schemas/Folder" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/PlanForbidden" }
        }
      },
      "post": {
        "tags": ["Folders"],
        "summary": "Create a folder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "description": "Folder name." },
                  "parent_id": { "type": "integer", "format": "int64", "description": "Parent folder. Omit to create at the root." }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "The created folder", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Folder" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/PlanForbidden" },
          "409": { "description": "A folder with that name already exists here", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Send your API key as a bearer token: `Authorization: Bearer trilli_sk_...`. Create keys on the API access page."
      }
    },
...
    "responses": {
      "Unauthorized": { "description": "Missing or invalid API key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "PlanForbidden": { "description": "API access is not included in this account's plan", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
...
      "Folder": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "tenant_id": { "type": "integer", "format": "int64" },
          "parent_folder_id": { "type": "integer", "format": "int64", "nullable": true },
          "name": { "type": "string" },
          "created_by_user_id": { "type": "integer", "format": "int64" },
          "status": { "type": "string", "example": "active" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  }
}